Start of hostname lookup.
Summary:
This is a starting point for looking up "real" DNS entries, even when they are
masked by cloudflare-enablement.

Right now it is specific to `haskell.org` but should be able to be generalized
fairly easily.

There are a lot of TODOs in Main.hs and this probably isn't ready to land, but
it's what I have so far.

Also `getZoneEntries` should probably move to the `CloudFlare` module or
something.

Test Plan:
```
ricky@t520 99% ~/devel/haskell/infra/cloudflare (cfhack)$ cloudflare www.haskell.org
(ip redacted)
```

Reviewers: austin

Reviewed By: austin

Differential Revision: https://phabricator.haskell.org/D378